+++ src/main/java/com/influxdb/v3/client/write/WriteOptions.java
@@ -40,7 +40,9 @@ import com.influxdb.v3.client.internal.Arguments;
  * <ul>
  *     <li><code>database</code> - specifies the database to be used for InfluxDB operations</li>
  *     <li><code>organization</code> - specifies the organization to be used for InfluxDB operations</li>
- *     <li><code>precision</code> - specifies the precision to use for the timestamp of points</li>
+ *     <li><code>precision</code> - specifies the precision to use for timestamps in line protocol records.
+ *     This setting is ignored when writing {@link com.influxdb.v3.client.Point}; for those writes, the client
+ *     always sends {@link WritePrecision#NS} precision to the server.</li>
  *     <li><code>defaultTags</code> - specifies tags to be added by default to all write operations using points.</li>
  *     <li><code>tagOrder</code> - specifies preferred tag order for point serialization.</li>
  *     <li><code>noSync</code> - skip waiting for WAL persistence on write</li>
@@ -124,6 +126,9 @@ public final class WriteOptions {
      *                      If it is not specified then use {@link ClientConfig#getDatabase()}.
      * @param precision     The precision to use for the timestamp of points.
      *                      If it is not specified then use {@link ClientConfig#getWritePrecision()}.
+     *                      This setting is ignored when writing {@link com.influxdb.v3.client.Point};
+     *                      for those writes, the client always sends {@link WritePrecision#NS}
+     *                      precision to the server.
      * @param gzipThreshold The threshold for compressing request body.
      *                      If it is not specified then use {@link WriteOptions#DEFAULT_GZIP_THRESHOLD}.
      */
@@ -140,6 +145,9 @@ public final class WriteOptions {
      *                      If it is not specified then use {@link ClientConfig#getDatabase()}.
      * @param precision     The precision to use for the timestamp of points.
      *                      If it is not specified then use {@link ClientConfig#getWritePrecision()}.
+     *                      This setting is ignored when writing {@link com.influxdb.v3.client.Point};
+     *                      for those writes, the client always sends {@link WritePrecision#NS}
+     *                      precision to the server.
      * @param gzipThreshold The threshold for compressing request body.
      *                      If it is not specified then use {@link WriteOptions#DEFAULT_GZIP_THRESHOLD}.
      * @param defaultTags   Default tags to be added when writing points.
@@ -158,6 +166,9 @@ public final class WriteOptions {
      *                      If it is not specified then use {@link ClientConfig#getDatabase()}.
      * @param precision     The precision to use for the timestamp of points.
      *                      If it is not specified then use {@link ClientConfig#getWritePrecision()}.
+     *                      This setting is ignored when writing {@link com.influxdb.v3.client.Point};
+     *                      for those writes, the client always sends {@link WritePrecision#NS}
+     *                      precision to the server.
      * @param gzipThreshold The threshold for compressing request body.
      *                      If it is not specified then use {@link WriteOptions#DEFAULT_GZIP_THRESHOLD}.
      * @param noSync        Skip waiting for WAL persistence on write.
@@ -187,6 +198,9 @@ public final class WriteOptions {
      *                      If it is not specified then use {@link ClientConfig#getDatabase()}.
      * @param precision     The precision to use for the timestamp of points.
      *                      If it is not specified then use {@link ClientConfig#getWritePrecision()}.
+     *                      This setting is ignored when writing {@link com.influxdb.v3.client.Point};
+     *                      for those writes, the client always sends {@link WritePrecision#NS}
+     *                      precision to the server.
      * @param gzipThreshold The threshold for compressing request body.
      *                      If it is not specified then use {@link WriteOptions#DEFAULT_GZIP_THRESHOLD}.
      * @param defaultTags   Default tags to be added when writing points.
@@ -209,6 +223,9 @@ public final class WriteOptions {
      *                      If it is not specified then use {@link ClientConfig#getDatabase()}.
      * @param precision     The precision to use for the timestamp of points.
      *                      If it is not specified then use {@link ClientConfig#getWritePrecision()}.
+     *                      This setting is ignored when writing {@link com.influxdb.v3.client.Point};
+     *                      for those writes, the client always sends {@link WritePrecision#NS}
+     *                      precision to the server.
      * @param gzipThreshold The threshold for compressing request body.
      *                      If it is not specified then use {@link WriteOptions#DEFAULT_GZIP_THRESHOLD}.
      * @param noSync        Skip waiting for WAL persistence on write.
@@ -234,6 +251,9 @@ public final class WriteOptions {
      *                      If it is not specified then use {@link ClientConfig#getDatabase()}.
      * @param precision     The precision to use for the timestamp of points.
      *                      If it is not specified then use {@link ClientConfig#getWritePrecision()}.
+     *                      This setting is ignored when writing {@link com.influxdb.v3.client.Point};
+     *                      for those writes, the client always sends {@link WritePrecision#NS}
+     *                      precision to the server.
      * @param gzipThreshold The threshold for compressing request body.
      *                      If it is not specified then use {@link WriteOptions#DEFAULT_GZIP_THRESHOLD}.
      * @param noSync        Skip waiting for WAL persistence on write.
@@ -265,6 +285,9 @@ public final class WriteOptions {
      *                      If it is not specified then use {@link ClientConfig#getDatabase()}.
      * @param precision     The precision to use for the timestamp of points.
      *                      If it is not specified then use {@link ClientConfig#getWritePrecision()}.
+     *                      This setting is ignored when writing {@link com.influxdb.v3.client.Point};
+     *                      for those writes, the client always sends {@link WritePrecision#NS}
+     *                      precision to the server.
      * @param gzipThreshold The threshold for compressing request body.
      *                      If it is not specified then use {@link WriteOptions#DEFAULT_GZIP_THRESHOLD}.
      * @param noSync        Skip waiting for WAL persistence on write.
@@ -307,6 +330,9 @@ public final class WriteOptions {
      *                      If it is not specified then use {@link ClientConfig#getDatabase()}.
      * @param precision     The precision to use for the timestamp of points.
      *                      If it is not specified then use {@link ClientConfig#getWritePrecision()}.
+     *                      This setting is ignored when writing {@link com.influxdb.v3.client.Point};
+     *                      for those writes, the client always sends {@link WritePrecision#NS}
+     *                      precision to the server.
      * @param gzipThreshold The threshold for compressing request body.
      *                      If it is not specified then use {@link WriteOptions#DEFAULT_GZIP_THRESHOLD}.
      * @param noSync        Skip waiting for WAL persistence on write.